refactor(telegram): isolate rich newline normalization
